Remove muiTheme context
