+++ js/src/ui/ContextProvider/contextProvider.js
@@ -24,14 +24,12 @@ import { LocaleStore } from '~/i18n';
 export default class ContextProvider extends Component {
   static propTypes = {
     api: PropTypes.object.isRequired,
-    muiTheme: PropTypes.object.isRequired,
     store: PropTypes.object,
     children: PropTypes.node.isRequired
   }
 
   static childContextTypes = {
     api: PropTypes.object,
-    muiTheme: PropTypes.object,
     store: PropTypes.object
   }
 
@@ -49,11 +47,10 @@ export default class ContextProvider extends Component {
   }
 
   getChildContext () {
-    const { api, muiTheme, store } = this.props;
+    const { api, store } = this.props;
 
     return {
       api,
-      muiTheme,
       store
     };
   }
