+++ js/src/ui/index.js
@@ -46,7 +46,6 @@ export Loading from './Loading';
 export MethodDecoding from './MethodDecoding';
 export { Busy as BusyStep, Completed as CompletedStep } from './Modal';
 export ModalBox from './ModalBox';
-export muiTheme from './Theme';
 export Page from './Page';
 export Popup from './Popup';
 export Portal from './Portal';
