+++ js/src/views/Playground/index.js
@@ -23,7 +23,6 @@ injectTapEventPlugin();
 
 import { initStore } from '@parity/shared/redux';
 import ContextProvider from '@parity/ui/ContextProvider';
-import muiTheme from '@parity/ui/Theme';
 
 import { api } from './parity';
 import Playground from './playground';
@@ -31,7 +30,7 @@ import Playground from './playground';
 const store = initStore(api, hashHistory);
 
 ReactDOM.render(
-  <ContextProvider api={ api } muiTheme={ muiTheme } store={ store }>
+  <ContextProvider api={ api } store={ store }>
     <Router history={ hashHistory }>
       <Route path='/' component={ Playground } />
     </Router>
