+++ js/src/views/Settings/Background/background.js
@@ -30,8 +30,7 @@ let counter = 0;
 
 class Background extends Component {
   static contextTypes = {
-    api: PropTypes.object.isRequired,
-    muiTheme: PropTypes.object.isRequired
+    api: PropTypes.object.isRequired
   }
 
   static propTypes = {
@@ -118,11 +117,9 @@ class Background extends Component {
   }
 
   onSelect = (seed) => {
-    const { muiTheme } = this.context;
     const { updateBackground } = this.props;
 
     return (event) => {
-      muiTheme.parity.setBackgroundSeed(seed);
       updateBackground(seed);
     };
   }
@@ -145,9 +142,9 @@ class Background extends Component {
   }
 
   generateSeed () {
-    const { api, muiTheme } = this.context;
+    const { api } = this.context;
 
-    return api.util.sha3.text(`${muiTheme.backgroundSeed}${Math.random()}${counter++}`);
+    return api.util.sha3.text(`${Math.random()}${counter++}`);
   }
 }
