+++ js/src/views/Settings/index.js
@@ -23,7 +23,6 @@ injectTapEventPlugin();
 
 import { initStore } from '@parity/shared/redux';
 import ContextProvider from '@parity/ui/ContextProvider';
-import muiTheme from '@parity/ui/Theme';
 
 import { api } from './parity';
 
@@ -36,7 +35,7 @@ import Settings from './settings';
 const store = initStore(api, hashHistory);
 
 ReactDOM.render(
-  <ContextProvider api={ api } muiTheme={ muiTheme } store={ store }>
+  <ContextProvider api={ api } store={ store }>
     <Router history={ hashHistory }>
       <Route path='/' component={ Settings }>
         <Route path='/background' component={ SettingsBackground } />
